Change dt from an Int to Float everywhere it was needed.
TileMap.h:

1) Changed int keyTime;
  int maxKeyTime; TO float keyTime;
  float maxKeyTime;

2) Changed TextureSelector(
   std::string texture_selector_data_file_path,
   float tile_size,
   float pos_x, float pos_y,
   float bounds_width, float bounds_height,
   const sf::Texture* texture_Sheet,
   sf::Font& hide_button_font,
   int key_time, int max_key_time TO  TextureSelector(
   std::string texture_selector_data_file_path,
   float tile_size,
   float pos_x, float pos_y,
   float bounds_width, float bounds_height,
   const sf::Texture* texture_Sheet,
   sf::Font& hide_button_font,
   float key_time, float max_key_time

TileMap.cpp:

1) Changed TILEMAP::TextureSelector::TextureSelector(
 std::string texture_selector_data_file_path,
 float tile_size,
 float pos_x, float pos_y,
 float bounds_width, float bounds_height,
 const sf::Texture* texture_Sheet,
 sf::Font& hide_button_font,
 int key_time, int max_key_time
) : keyTime(key_time), maxKeyTime(max_key_time) TO TILEMAP::TextureSelector::TextureSelector(
 std::string texture_selector_data_file_path,
 float tile_size,
 float pos_x, float pos_y,
 float bounds_width, float bounds_height,
 const sf::Texture* texture_Sheet,
 sf::Font& hide_button_font,
 float key_time, float max_key_time
) : keyTime(key_time), maxKeyTime(max_key_time)

2) Changed   this->keyTime += static_cast<int>(2.f * dt * (1.f / dt)); TO   this->keyTime += 2.f * dt * (1.f / dt); IN /*Update Functions*/
void TILEMAP::TextureSelector::updateKeyTime(const float& dt)

State.h:

1) Changed  int keyTime;
 int maxKeyTime; TO float keyTime;
 float maxKeyTime;

2) Changed  this->keyTime = 0;
 this->maxKeyTime = 25; TO this->keyTime = 0.f;
 this->maxKeyTime = 25.f; IN /*Constuctor & Destructor*/
State::State(GameInfo* game_info)

3) Changed   this->keyTime += 2 * static_cast<int>(dt * (1.f / dt)); TO  this->keyTime += 2 * dt * (1.f / dt); IN void State::updateKeyTime(const float& dt)

GUI.h:

1) Changed int keyTime;
  int maxKeyTime; TO float keyTime;
  float maxKeyTime;

2) Changed /*Constuctor / Destructor*/
  DropdownList(
   float pos_x, float pos_y,
   float width, float height,
   sf::Font& font, std::string list[], unsigned int character_size,
   int& key_time, int& max_key_time,
   unsigned numofElements, unsigned default_index = 0 TO /*Constuctor / Destructor*/
  DropdownList(
   float pos_x, float pos_y,
   float width, float height,
   sf::Font& font, std::string list[], unsigned int character_size,
   float& key_time, float& max_key_time,
   unsigned numofElements, unsigned default_index = 0

GUI.cpp:

1) Changed /*Constructor/Destructor*/
GUI::DropdownList::DropdownList(
 float pos_x, float pos_y,
 float width, float height,
 sf::Font& font_reference, std::string list[], unsigned int character_size,
 int& key_time, int& max_key_time,
 unsigned numofElements, unsigned default_index TO /*Constructor/Destructor*/
GUI::DropdownList::DropdownList(
 float pos_x, float pos_y,
 float width, float height,
 sf::Font& font_reference, std::string list[], unsigned int character_size,
 float& key_time, float& max_key_time,
 unsigned numofElements, unsigned default_index

+++ Game Engine/GUI.cpp
@@ -129,7 +129,7 @@ GUI::DropdownList::DropdownList(
 	float pos_x, float pos_y,
 	float width, float height,
 	sf::Font& font_reference, std::string list[], unsigned int character_size,
-	int& key_time, int& max_key_time,
+	float& key_time, float& max_key_time,
 	unsigned numofElements, unsigned default_index
 )
 	: font(font_reference), showList(false), keyTime(key_time), maxKeyTime(max_key_time)
