+++ Game Engine/State.cpp
@@ -19,8 +19,8 @@ State::State(GameInfo* game_info)
 	this->window = game_info->window;
 	this->sfmlEvent = game_info->sfmlEvent;
 	this->supportedKeys = game_info->supportedKeys;
-	this->keyTime = 0;
-	this->maxKeyTime = 25;
+	this->keyTime = 0.f;
+	this->maxKeyTime = 25.f;
 	this->isQuit = false;
 	this->mouseReleased = false;
 	this->tileSize = 32.f;
@@ -115,7 +115,7 @@ void State::updateMousePosition(sf::View* view, sf::View* default_window_view)
 void State::updateKeyTime(const float& dt)
 {
 	if (this->keyTime < this->maxKeyTime)
-		this->keyTime += 2 * static_cast<int>(dt * (1.f / dt));
+		this->keyTime += 2 * dt * (1.f / dt);
 
 	/*Test Debug
 	std::cout << this->keyTime << '\n';
